+++ sqlalchemy_things/__init__.py
@@ -1,4 +1,4 @@
 from sqlalchemy_things import column_types, declarative, pagination
 
-__version__ = '0.9.1'
+__version__ = '0.10.0'
 __all__ = ['declarative', 'column_types', 'pagination']